Latest version of OSeMOSYS short code, v2016_08_01
Compared to the standard version, the short version has a more synthetic structure: many variables replaced in the equations by others they derive from, in such way that the number of equations is highly reduced. For this reason, the user will see in the file  many variables and equations commented out (through #). Whenever the users modify the code, tough, they are recommended to modify accordingly all the equations, also those commented.