Added Cascaded Hydro Modelling Code
This Hydro modelling code is based on the original OSeMOSYS storage equations but is designed to make the modelling of cascaded hydro facilities more efficient and easier to follow.  A new set (RESERVOIR) is created that is similar to the STORAGE set that exists in original OSeMOSYS.  Dam technologies and river inflows are defined for each RESERVOIR and allow modelling using intuitive units of inflows and power conversion.