Add link to the original OSeMOSYS paper
To newcomers to the project OSeMOSYS, the original paper
is a good introduction to the structure of the model.

The code described in the paper is slighlt outdated by now but
the underlying model has stayed mostly the same and the
paper is a good resource for people wanting to have a deeper
understanding of the model.

One argument against having the link as part of the README
is that it needs a paid access to ScienceDirect. But I think that
most of the modeling community has ways to access it through
their institutions so maybe it's not an issue?